split CheckMojoTest
